+++ Modules/Intent.Modules.ModuleBuilder/Templates/Api/ApiElementExtensionModel/ApiElementExtensionModelTemplateRegistration.cs
@@ -26,6 +26,7 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementExtensionModel
 
         public override string TemplateId => ApiElementExtensionModelTemplate.TemplateId;
 
+        [IntentManaged(Mode.Fully)]
         public override ITemplate CreateTemplateInstance(IOutputTarget outputTarget, ElementExtensionModel model)
         {
             return new ApiElementExtensionModelTemplate(outputTarget, model);