+++ Modules/Intent.Modules.ModuleBuilder/Templates/Api/ApiPackageExtensionModel/ApiPackageExtensionModelTemplateRegistration.cs
@@ -26,6 +26,7 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiPackageExtensionModel
 
         public override string TemplateId => ApiPackageExtensionModelTemplate.TemplateId;
 
+        [IntentManaged(Mode.Fully)]
         public override ITemplate CreateTemplateInstance(IOutputTarget outputTarget, PackageExtensionModel model)
         {
             return new ApiPackageExtensionModelTemplate(outputTarget, model);