+++ Modules/Intent.Modules.ModuleBuilder/Templates/TemplateExtensions/TemplateExtensionsTemplateRegistration.cs
@@ -10,8 +10,8 @@ using Intent.Modules.ModuleBuilder.Templates.Registration.FilePerModel;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
-[assembly: IntentTemplate("Intent.ModuleBuilder.TemplateRegistration.SingleFileNoModel", Version = "1.0")]
 [assembly: DefaultIntentManaged(Mode.Merge)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.TemplateRegistration.SingleFileNoModel", Version = "1.0")]
 
 namespace Intent.Modules.ModuleBuilder.Templates.TemplateExtensions
 {