+++ Modules/Intent.Modules.ModuleBuilder/IntentNugetPackages.cs
@@ -9,7 +9,7 @@ namespace Intent.Modules.ModuleBuilder
         public static NugetPackageInfo IntentPackager = new NugetPackageInfo("Intent.Packager", "3.4.0")
             .SpecifyAssetsBehaviour(new[] { "all" }, new[] { "runtime", "build", "native", "contentfiles", "analyzers", "buildtransitive" });
 
-        public static NugetPackageInfo IntentModulesCommon = new("Intent.Modules.Common", "3.4.1");
+        public static NugetPackageInfo IntentModulesCommon = new("Intent.Modules.Common", "3.4.3");
         public static NugetPackageInfo IntentModulesCommonTypes = new("Intent.Modules.Common.Types", "3.3.9");
     }
 }